Properly fixed test signatures, added signatures for actual LIC conditiion functions

+++ src/lab1/test/LaunchInterceptorConditionsTest.java
@@ -22,7 +22,7 @@ public class LaunchInterceptorConditionsTest {
         double[] y = new double[] {0, 0, 0, 0, 0, 0, 0};
         int c_pts = 2;
         int d_pts = 2;
-        int epsilon = 1;
+        double epsilon = 1;
         boolean result = LaunchInterceptorConditions.condition9(x, y, c_pts, d_pts, epsilon, 7);
         assertEquals(result, false);
     }
@@ -36,7 +36,7 @@ public class LaunchInterceptorConditionsTest {
         double[] y = new double[] {0, 0, 0, 0, 0, 0, 0};
         int c_pts = 2;
         int d_pts = 2;
-        int epsilon = 1;
+        double epsilon = 1;
         boolean result = LaunchInterceptorConditions.condition9(x, y, c_pts, d_pts, epsilon, 7);
         assertEquals(result, false);
     }
@@ -50,7 +50,7 @@ public class LaunchInterceptorConditionsTest {
         double[] y = new double[] {0, 0, 0, 0};
         int c_pts = 1;
         int d_pts = 1;
-        int epsilon = 1;
+        double epsilon = 1;
         boolean result = LaunchInterceptorConditions.condition9(x, y, c_pts, d_pts, epsilon, 4);
         assertEquals(result, false);
     }
@@ -65,7 +65,7 @@ public class LaunchInterceptorConditionsTest {
         double[] y = new double[] {0, 2, 3, 5, 4, 5, 0};
         int e_pts = 2;
         int f_pts = 2;
-        int area1 = 0;
+        double area1 = 0;
         boolean result = LaunchInterceptorConditions.condition10(x, y, e_pts, f_pts, area1, 7);
         assertEquals(result, true);
     }
@@ -79,7 +79,7 @@ public class LaunchInterceptorConditionsTest {
         double[] y = new double[] {0, 0, 0, 0};
         int e_pts = 0;
         int f_pts = 0;
-        int area1 = 0;
+        double area1 = 0;
         boolean result = LaunchInterceptorConditions.condition10(x, y, e_pts, f_pts, area1, 4);
         assertEquals(result, false);
     }
